pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "finfluencer-tracker"
version = "1.0.0"
description = "Track and verify predictions from Indian finance YouTubers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "youtube-transcript-api>=1.2.0",
    "yt-dlp>=2024.1.0",
    "google-generativeai>=0.5.0",
    "nselib>=1.0.0",
    "exa-py>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "sqlite-utils>=3.35",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "click>=8.1.0",
    "rich>=13.7.0",
]

[project.scripts]
finfluencer = "main:cli"
